requirements.txt
django==3.1.4
djangorestframework==3.12.2
pillow
django-admin-interface
django-crispy-forms
django-rest-framework
social-auth-app-django
django-pwa